requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
Deprecated==1.2.13
Django==4.0.4
djangorestframework==3.13.1
GPUtil==1.4.0
packaging==21.3
psutil==5.9.0
pyparsing==3.0.8
pytz==2022.1
redis==4.2.2
sqlparse==0.4.2
wrapt==1.14.0
